pyproject.toml
[project]
name = "pytutor-llms-suede"
version = "0.1.0"
description = ""
authors = [
    { name = "Github Action (authored by pmalacho-mit)", email = "pmalacho-mit@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "google-genai (>=1.66.0,<2.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "anthropic (>=0.84.0,<0.85.0)",
    "instructor (>=1.14.5,<2.0.0)",
    "datamodel-code-generator (>=0.54.1,<0.55.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "jsonref (>=1.1.0,<2.0.0)",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"
asyncio_mode = "auto"
markers = [
    "benchmark: long-running live provider benchmark tests",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "ipykernel (>=7.2.0,<8.0.0)",
    "ipython (>=9.11.0,<10.0.0)",
]

[tool.poetry.group.dev]
optional = true
